Fix for bug #25048: eeprom.h will not compile as c++ code. Undo (partially) the rev 1.27 (before 2008-08-19), eeprom_read_byte(): cast operation of address into 16-bit integer regardless of EEPROM size. Change (this is cosmetic) the (unsigned) cast to (size_t) cast in both byte functions.
git-svn-id: https://svn.savannah.nongnu.org/svn/avr-libc/trunk/avr-libc@1819 1084037a-409f-4a99-8160-1a4b3577d4c1

+++ include/avr/eeprom.h
@@ -196,9 +196,9 @@ __ATTR_PURE__ static __inline__ uint8_t eeprom_read_byte (const uint8_t *__p)
 {
     do {} while (!eeprom_is_ready ());
 #if E2END <= 0xFF
-    EEARL = (uint8_t)__p;
+    EEARL = (size_t)__p;
 #else
-    EEAR = (uint16_t)__p;
+    EEAR = (size_t)__p;
 #endif
     /* Use inline assembly below as some AVRs have problems with accessing
     EECR with STS instructions. For example, see errata for ATmega64. 
@@ -306,9 +306,9 @@ static __inline__ void eeprom_write_byte (uint8_t *__p, uint8_t __value)
 #endif
 
 #if	E2END <= 0xFF
-    EEARL = (unsigned)__p;
+    EEARL = (size_t)__p;
 #else
-    EEAR = (unsigned)__p;
+    EEAR = (size_t)__p;
 #endif
     EEDR = __value;
